make FlightRecorderConnection an interface
