+++ jfr-connection/src/main/java/io/opentelemetry/contrib/jfr/connection/PredefinedConfiguration.java
@@ -29,7 +29,7 @@ public class PredefinedConfiguration implements RecordingConfiguration {
       String[] argTypes = new String[] {long.class.getName(), String.class.getName()};
       mBeanServerConnection.invoke(objectName, "setPredefinedConfiguration", args, argTypes);
     } catch (InstanceNotFoundException | MBeanException | ReflectionException e) {
-      throw FlightRecorderConnection.canonicalJfrConnectionException(
+      throw JfrConnectionException.canonicalJfrConnectionException(
           getClass(), "invokeSetConfiguration", e);
     }
   }
